time: replace uint32_t with int64_t for Bitcoin Core time parity
Replaces uint32_t timestamps with int64_t throughout AddressManager to
match Bitcoin Core and prevent 2106 overflow. Removes now() wrapper
function in favor of direct util::GetTime() calls.

Changes:
- AddrInfo struct: Changed timestamp, last_try, last_success from uint32_t to int64_t
- Time constants: Changed SECONDS_PER_DAY, RECENT_TRY_SEC, ADDRMAN_HORIZON, ADDRMAN_MIN_FAIL to int64_t
- Method signatures: Updated all time-related functions to use int64_t
- Removed now() wrapper: Direct use of util::GetTime() (already returns int64_t)
- Persistence: Fixed Load() to deserialize timestamps as int64_t (was uint32_t, causing truncation)
- Protocol compatibility: Cast to uint32_t when creating TimestampedAddress for wire protocol
- Tests: Updated timestamp variables from uint32_t to int64_t

Benefits:
✅ No 2106 overflow (int64_t valid until year 292 billion)
✅ Bitcoin Core parity (NodeSeconds uses std::chrono with seconds precision)
✅ Consistent time handling throughout AddressManager
✅ Simpler code (removed wrapper function)
✅ Proper persistence (no timestamp truncation on load/save)

Note: protocol::TimestampedAddress still uses uint32_t for wire protocol
compatibility. Timestamps are cast to uint32_t when serializing for
network transmission (get_addresses() method).

+++ test/network/dos/addr_echo_suppression_tests.cpp
@@ -14,7 +14,7 @@ using namespace coinbasechain;
 using namespace coinbasechain::test;
 using namespace coinbasechain::protocol;
 
-static protocol::TimestampedAddress MakeTsAddrIPv4(const std::string& ip_v4, uint16_t port, uint32_t ts) {
+static protocol::TimestampedAddress MakeTsAddrIPv4(const std::string& ip_v4, uint16_t port, int64_t ts) {
     protocol::TimestampedAddress ta;
     ta.timestamp = ts;
     ta.address.services = protocol::ServiceFlags::NODE_NETWORK;
@@ -48,7 +48,7 @@ TEST_CASE("Echo suppression: node does not echo addresses learned from same peer
     REQUIRE(orch.WaitForConnection(A, B));
 
     // B announces address X to A via ADDR
-    const uint32_t ts = static_cast<uint32_t>(coinbasechain::util::GetTime());
+    const int64_t ts = coinbasechain::util::GetTime();
     auto X = MakeTsAddrIPv4("10.0.0.42", ports::REGTEST, ts);
     message::AddrMessage addr_msg; addr_msg.addresses.push_back(X);
     auto payload = addr_msg.serialize();
@@ -92,7 +92,7 @@ TEST_CASE("Echo suppression is per-peer: addresses from C are served to other pe
     REQUIRE(orch.WaitForConnection(A, C));
 
     // C announces X to A
-    const uint32_t ts = static_cast<uint32_t>(coinbasechain::util::GetTime());
+    const int64_t ts = coinbasechain::util::GetTime();
     auto X = MakeTsAddrIPv4("10.0.0.99", ports::REGTEST, ts);
     message::AddrMessage addr_msg; addr_msg.addresses.push_back(X);
     auto payload = addr_msg.serialize();
@@ -152,7 +152,7 @@ TEST_CASE("Echo suppression TTL expiry allows address to be served back after 10
     REQUIRE(orch.WaitForConnection(A, B));
 
     // B announces Y to A
-    const uint32_t ts2 = static_cast<uint32_t>(coinbasechain::util::GetTime());
+    const int64_t ts2 = coinbasechain::util::GetTime();
     auto Y = MakeTsAddrIPv4("10.0.0.77", ports::REGTEST, ts2);
     message::AddrMessage addr_msg2; addr_msg2.addresses.push_back(Y);
     auto payload2 = addr_msg2.serialize();
